scx_rustland_core: prevent deadlock with per-CPU DSQs and CPU affinity
If a task that is executing sched_setaffinity() is dispatched on a
per-CPU DSQ it may stall the DSQ completely, since the task won't be
able to be consumed from the corresponding CPU.

This can be easily triggered running the following stress test:

  $ stress-ng --aggressive -c (nproc) -f (nproc)

From the stall trace we can see something like the following:

  R stress-ng[2648662] -6880ms
      scx_state/flags=3/0x9 dsq_flags=0x1 ops_state/qseq=0/0
      sticky/holding_cpu=-1/-1 dsq_id=0x5 dsq_vtime=0
      cpus=ff

    __set_cpus_allowed_ptr+0x1c8/0x260
    __sched_setaffinity+0x105/0x1c0
    sched_setaffinity+0x1ed/0x2d0
    __x64_sys_sched_setaffinity+0xa5/0x100
    do_syscall_64+0x82/0x190
    entry_SYSCALL_64_after_hwframe+0x76/0x7e

This should probably be addressed in the core sched_ext, but for now
prevent this deadlock by tracking when a task is executing
sched_setaffinity() and automatically bounce those tasks to the shared
DSQ (that can be consumed from any CPU).

This should solve all the recent CI failures with the scx_rustland_core
schedulers.

Signed-off-by: Andrea Righi <andrea.righi@linux.dev>