scx_layered: min_exec_us should only affect vtime delta
min_exec_us is used to penalize tasks which switch frequently in competition
with other tasks in the same layer. This shouldn't affect costc or CPU
consumption accounting. In the current code, it's easy to get non-sensical
util numbers with e.g. `stress-ng --workload` which then affects CPU
allocations in unintended ways. Fix it.