scx_layered: Reimplement refresh_cpumasks() using the calculated target CPU allocations
Currently, the layer growth and shrinking are implemented locally within
each layer and allocation decisions are made without knowledge of global
state which makes it impossible e.g. for a layer to know when it needs to
shrink due to compeition from other layers.

A preceding commit implmented weight-based CPU allocation calculation, so we
now know what the eventual distribution should look like. Reimplement
refresh_cpumasks() accordingly so that CPUs are allocated to the layers
according to their weights.