scx_layered: Collect and report per-llc-layer stats
Collect and report how many tasks are being scheduled on each llc-layer pair
along with running avg of queueing latency. This will be used to affect
scheduling behaviors in the future.

Reading stats from rust would be a lot easier if llc_ctx's were in a bss
array but putting in a bss array triggers verifier bugs, so do the ugly
slice dancing. Convert to bss array once we can ignore broken kernels. See
the comment in BpfStats::read().