scx_layered: Implement per-task avg runtime tracking and use it to calculate q latency
Per-LLC layer queueing latencies were measured on each ops.running()
transition and runtime avarged. Depending on the specific task, this average
can swing wildly and it's difficult to base scheduling decisions on them.

Instead, track per-task average runtime and then use them to determine the q
latency as the sum of the average runtimes of the tasks on the q. While this
requires atomic ops to maintain the sum, the operations are mostly LLC local
and not noticeable. The up-to-date information will help making better
scheduling decisions which should more than offset whatever additional
overhead.

LLC_LSTAT_LAT is now only used to monitor how llc_ctx->queued_runtime is
behaving, so decay it slower. Also, don't quelch it to zero when
LLC_LSTAT_CNT is 0 so that bugs in queued_runtime maintenance are visible.