scx_layered: Compensate for layer usage update fluctuations
layer_usages are updated at ops.stopping(). If tasks in a layer keep
running, it may not update the usage stats for a long time to the point
where the reported usage stats fluctuate wildly making CPU allocation
oscillate. Compensate for it by adding the time spent for the currently
running task.