scx_stats,scx_lavd: describe non-nested stats meta
In cases where non-nested structs are used for statistics (such as in scx_lavd), the describe_metadata method can now display metadata for all of them.
In scx_lavd, two distinct monitoring modes rely on separate structs. Therefore, it was necessary to find a way that accommodates both these separate statistics.

The issue is reported in sched-ext#783

Signed-off-by: Luigi De Matteis <ldematteis123@gmail.com>