scx_layered: Use low fallback DSQs for empty layers
Now that low fallback DSQs are usable, we can put empty layer tasks there
instead of depending on the fallback CPU. The fallback CPU is still needed
because race conditions can put stray tasks in empty layers but this will be
rare and should be negligible.

As the fallback CPU was already prioritizing empty layers after only the hi
fallbac, the owned protection wasn't doing much and is completely
unnecessary now. Rip it out.

Also add fallback CPU util metric to verify that it's always staying close
to zero.