scx_layered: Implement empty LLC draining
A layer is served by per-LLC DSQs. Only tasks that can be serviced by all
CPUs in an LLC are put in the DSQ, so as long as a CPU is assigned to the
layer in the LLC, forward progress is guaranteed. However, if a layer-LLC
loses all its CPUs, there is no forward progress guarantee ignoring the
antistall mechanism. For a confined layer, no CPU will be visiting the empty
LLCs and even for a grouped or open layer, execution from an LLC without
CPUs assigned is lower priority than owned execution and can easily starve.

To resolve the problem, implement LLC draining mechanism. When layer-LLC
loses all CPUs with tasks in it, draining is turned on and other CPUs
assigned to the layer will alternate between their own execution and
draining LLCs without any CPU. The interlockings between the involved code
paths - refresh_cpumasks(), layered_enqueue() and layered_dispatch() - are
rather intricate to guarantee that no tasks end up sitting in a CPU-less
LLC. See comments for details.