scx_layered: Fix queued_runtime tracking
When a task gets dequeued and its property is changed so that it belongs to
a different layer and then requeued, the current qrt tracking code deducts
its runtime_avg from the wrong layer which leads to e.g. qrt underflows.
Remember the layer the runtime is charged to and deduct it from there.