scx_rusty: prevent scheduling bubbles on kernels >= 6.12
Selecting an idle CPU without dispatching a task to it can lead to
scheduling bubbles, as the CPU remains incorrectly marked as busy and is
excluded from the ops.select_cpu() selection, resulting in inefficient
core utilization.

To prevent this, always check the idle state of the CPUs that we want to
wake up, but avoid marking them as busy, unless we are certain that a
task will be dispatched on them.

Note that this issue occurs only in kernel 6.12 and later (following the
pick_next_task() refactoring), as the in-kernel CPU idle state update
happens only during transitions between a non-idle task and the idle
task, whereas earlier kernels refreshed the CPU idle state on every idle
task cycle (as a result of the put_prev_task() call).

Signed-off-by: Andrea Righi <arighi@nvidia.com>