scx_bpfland: Set max performance level with balance_performance
The "balance_performance" energy profile is intended to provide a middle
ground between power savings and performance. However, when running with
this profile, there is no strong reason to enforce power-saving
measures, especially in scenarios where performance is the primary
concern.

For laptops, this choice is particularly relevant because
balance_performance is typically used when the system is plugged into AC
power. In this case, restricting power consumption is unnecessary, as
battery life is not a constraint. Instead, we should prioritize higher
performance for better responsiveness and throughput.

Therefore, use the maximum performance level with the
balance_performance energy profile.

Signed-off-by: Andrea Righi <arighi@nvidia.com>