scx_lavd: Conder the hardware resource sharing level in CPU ordering
When determining the preference ordering of CPUs for performance
and power-save mode, consider a CPU's hardware resource sharing levels
-- such as whether a CPU shares an L2/L3 cache or a physical core.
For the performance mode, in which using an SMT core is not preferred,
CPUs that share fewer hardware resources are preferred. On the other hand,
for the power-save mode, in which using an SMT core is preferred, CPUs
that share more hardware resources are preferred.

Signed-off-by: Changwoo Min <changwoo@igalia.com>