scx_rusty: Add task_ctx.pid for debug visibility
BPF and userland are now identifying tasks with pointers to task_ctx.
However, it's difficult to know which task that maps to from userspace. Add
task_ctx.pid so that userspace can map task_ctx to the task. A trace
printout is added on each migration verdict for visibility.

Note that a task's pid may change during multithreaded execs and
task_ctx.pid may go out of sync in such cases. We can update this
dynamically but as this is just for debug visibility, leave it be for now.