scx_lavd: Use BPF CO-RE for SCX_ENQ_CPU_SELECTED
There could be a case such that the scheduler is compiled against the latest
vmlinux.h, which has SCX_ENQ_CPU_SELECTED, but runs on the old kernel, which
does not have SCX_ENQ_CPU_SELECTED. In this case, !(enq_flags &
SCX_ENQ_CPU_SELECTED) will be always true, which is unintended. So use BPF
CO-RE to check if SCX_ENQ_CPU_SELECTED exists in runtime.

Signed-off-by: Changwoo Min <changwoo@igalia.com>