ci: enable running on github runners when outside of repository
The build Nix kernel jobs currently fail in external forks because they expect
access to the self-hosted runner. Change this to dynamically selecting the
self-hosted runner when running in a repo owned by "sched-ext" (the only repos
with access to our runner) and "ubuntu-latest" otherwise. Install Nix on GitHub
runners but not on self-hosted runners (assume it's already there for now).

This involves

Test plan:
- Pushed to my fork and (after re-enabling actions) both builds passed.
- Runs in the CI on this PR. Will check it hits the right machine.