ci: pass through rust test paths to prevent a rebuild
This is a fun one that's been bothering me for a while. The Rust test
jobs in the CI have been completely rebuilding the test targets in the
VM really slowly even though they are built before the vng VM is entered.

I thought this was something to do with the virtual filesystem before,
but it turned out to be much more mundane. The caches for rustup and
cargo are stored in $HOME/.rustup and $HOME/.cargo, and vng runs under a
different user in the VM (root I assume) so is looking in the wrong
place.

Pass through `RUSTUP_HOME` and `CARGO_HOME` to the `cargo test` call in
the VM. This prevents the re-sync and download of the rustup components
and prevents the rebuild/redownload of the Rust crates.

This makes the performance of the CI drastically better.

Test plan:
- CI