scxtop: track number of dropped events and render it
Adds a stat to track events which can't be inserted into the ringbuf. I can't
get this to hit on my big AMD machines which is a good sign. I've tried a
couple of prod services and some stress-ng configurations. Seems like dropping
events isn't an issue at the minute, but we should keep an eye on it.

There are likely better ways to render this, but at least this way we keep
track of it and can render it more appropriately in the future.

Test plan:
- Artificially added failures on the BPF side. They render as expected in
  `scxtop` during a trace.