scxtop: run clippy and implement fixes
Ran `clippy` against `scxtop` and kept iterating until no more warnings. I was
very happy with the quality of these warnings.

The main change to the code is things are less indented because of some
restructuring. This is another PR to review with GitHub set to not show
whitespace changes.

Test plan:
- Build scxtop and run any tests we write in the CI in the future.
- Run `cargo clippy` only for scxtop in the CI too to keep things clean.