scxtop: add experimental_long_tail_tracing
Adding an experimental feature to automatically start a perfetto trace when a
named function takes a specified long time to return. This can be used to get a
trace of an event when a group of long tail events happen together.

Prefacing this all with experimental as the implementation is a little funky.

Implementation notes:
- Starts the trace entirely from BPF space and relies on ordering of the
  ringbuffer to make sure userspace notices before the first events come
  through.
- Misses early softirq events because they require userspace to attach the
  programs.
- Starts the trace from `trace_tick_warmup` ticks to events aren't dropped.
- Extends the trace if another such event comes from, meaning this can generate
  an infinitely long trace and probably get scxtop OOM killed if you get the
  parameters wrong.

The concept here is that many traces can be recorded by leaving scxtop running,
but currently it doesn't seem to be able to get more than 2 without the TUI
crashing and leaving it in a terminal output that needs Ctrl+C'ing.

Test plan:
- Attached to a production binary and it hits when expected, gathering traces
  that look reasonable.